+++ NEWS.md
@@ -6,6 +6,7 @@ Improvements
 
 Bug fixes 
 * Fix issue where `*_structure` functions did not correctly parse strings containing quotes in some cases (#466)
+* Fix issue where page_plan doesn't work if there is only one level in your paging variable. 
 
 # tfrmt 0.1.3